Edit Delete Post clean up (#146)
* #135 Edit Delete post cleanup

* #135 Fix comment.js indentation

* #146 Use string concatenation for post.id

* #146 Seperate Post Card and Form

* #146 Make Author Profile page posts chronological
