--- code/socialDistribution/static/js/comment.js
@@ -1,6 +1,0 @@
-    /**
-     * Redirect to home page of author
-     * */
-     function redirectToAuthorPage(id) {
-        location.href = `${location.protocol}//${location.host}/app/author/${id}`;
-    }
