+++ code/socialDistribution/static/js/comment/comment.js
@@ -1,0 +1,6 @@
+/**
+ * Redirect to home page of author
+ * */
+function redirectToAuthorPage(id) {
+    location.href = `${location.protocol}//${location.host}/app/author/${id}`;
+}
