+++ code/socialDistribution/templates/tagtemplates/comment.html
@@ -62,4 +62,4 @@
 </div>
 
 {% load static %}
-    <script src="{% static 'js/comment.js' %}"></script>
+    <script src="{% static 'js/comment/comment.js' %}"></script>
