+++ code/socialDistribution/templates/tagtemplates/modal.html
@@ -1,5 +1,5 @@
 
-{% load post %}
+{% load post_form %}
 <div class="modal fade" id="{{modal_id}}" tabindex="-1" aria-labelledby="{{modal_label}}" aria-hidden="true">
     <div class="modal-dialog" role="document">
         <div class="modal-content">
