+++ code/socialDistribution/templates/tagtemplates/post.html
@@ -48,16 +48,36 @@
                 {% if isAuthor %}
                 <span>
                     <!-- Launch Edit Modal -->
-                    <a href="{% url 'socialDistribution:commentPost' post.id %}?action=edit"> <button
-                            class="btn btn-secondary"><i class="fas fa-edit"></i></button></a>
+                    <button 
+                        type="button" 
+                        class="btn btn btn-secondary" 
+                        data-bs-toggle="modal"
+                        data-bs-target="#edit_post_modal{{ post.id }}">
+                            <i class="fas fa-edit"></i>
                     </button>
+           
                     <!-- Launch Delete Modal -->
-                    <a href="{% url 'socialDistribution:commentPost' post.id %}?action=delete"> <button
-                            class="btn btn btn-danger"><i class="fas fa-trash-alt"></i></button></a>
+                    <button 
+                        type="button" 
+                        class="btn btn btn-danger" 
+                        data-bs-toggle="modal" 
+                        data-bs-target="#delete_post_modal{{ post.id }}">
+                            <i class="fas fa-trash-alt"></i>
                     </button>
                 </span>
                 {% endif %}
             </div>
         </div>
     </div>
+
+    {% load modal %}
+    {% load utility %}
+    <form method="POST" action="{% url 'socialDistribution:editPost' post.id %}">
+        {% csrf_token %}
+        {% modal id="edit_post_modal"|addstr:post.id type="edit" label="Edit Post" title="Edit Post" btn="Save" postid=post.id %}
+    </form>
+    <form method="POST" action="{% url 'socialDistribution:deletePost' post.id %}">
+        {% csrf_token %}
+        {% modal id="delete_post_modal"|addstr:post.id type="delete" label="Delete Post" title="Delete Post" btn="Delete" postid=post.id%}
+    </form>
 </div>
